requirements.txt
openai~=0.27.8
python-dotenv~=1.0.0
diskcache~=5.6.1
parameterized~=0.9.0